split up the stories to be better examples
